h_approve_knock: fix D1_TYPE_ERROR from binding i64 as BigInt
Found live, on the first real end-to-end test of the human loop
(model knocks, founder approves on the real dashboard, model uses the
returned token) -- every prior verification of this path used a
synthetic D1 row inserted directly via wrangler d1 execute, never the
actual INSERT this handler runs.

expires_at and now are i64 unix seconds; `.into()` on an i64 produces
a JS BigInt via wasm-bindgen's default conversion, and D1's bind()
rejects BigInt outright: "D1_TYPE_ERROR: Type 'bigint' not supported
for value '...'". The knock still decided and logged as approved --
that happens inside decide_knock, before this code ever runs -- so the
operator saw "knock.approved" in the activity log while the actual
capability silently failed to write. A real partial-failure state with
no visible signal beyond a 500 the frontend didn't surface loudly.

Fixed by casting through f64 before binding, matching the convention
scratchpad.rs already uses everywhere else in this codebase for the
exact same reason -- this was the one D1 insert site that didn't.

Also turned on observability (logs, invocation_logs) on curatom-kernel
via the Cloudflare API to actually see this error, since it previously
had none -- unlike hostos-mcp, which already logs.

+++ workers/api/src/lib.rs
@@ -654,12 +654,19 @@ impl CuratomKernel {
                    scope = excluded.scope, expires_at = excluded.expires_at, \
                    revoked = 0, knock_id = excluded.knock_id",
             );
+            // D1's bind rejects a JS BigInt outright -- `worker`'s
+            // `From<i64> for JsValue` produces one, since i64 doesn't fit an
+            // f64 losslessly in general. A unix-seconds timestamp does (well
+            // under 2^53), so cast through f64 rather than i64 -> JsValue
+            // directly. Verified live: the raw i64 bind failed with
+            // "D1_TYPE_ERROR: Type 'bigint' not supported" on the very first
+            // real approval this was exercised against.
             let bound = match stmt.bind(&[
                 key_hash.clone().into(),
-                expires_at.into(),
+                (expires_at as f64).into(),
                 knock.name.clone().into(),
                 knock.id.clone().into(),
-                now.into(),
+                (now as f64).into(),
             ]) {
                 Ok(b) => b,
                 Err(e) => return (500, json!({ "error": format!("bind: {e}") })),
